rosc
Module encoder

Encodes an OscPacket to a byte vector.

Structs

WriteOutput

A newtype which can be used to wrap any type which implements std::io::Seek and std::io::Write to allow it to be used as an Output.

Traits

Output

A trait for values that can receive encoded OSC output via encode_into. This allows more flexibility in how the output is handled, including reusing part of an existing buffer or writing directly to an external sink (e.g. a file).

Functions

encode

Takes a reference to an OSC packet and returns a byte vector on success. If the packet was invalid an OscError is returned.

encode_into

Takes a reference to an OSC packet and writes the encoded bytes to the given output. On success, the number of bytes written will be returned. If an error occurs during encoding, encoding will stop and the error will be returned. Note that in that case, the output may have been partially written!

encode_string

Null terminates the byte representation of string s and adds null bytes until the length of the result is a multiple of 4.

encode_string_into

Writes the given string s to the given Output, adding 1-4 null bytes such that the length of the result is a multiple of 4.

pad

Returns the position padded to 4 bytes.