Crate rosc

rosc is an implementation of the OSC 1.0 protocol in pure Rust.

Modules

decoder

Provides a decoding method for OSC packets.

encoder

Encodes an OscPacket to a byte vector.

Structs

OscArray

An OscArray color.

OscBundle

An OSC bundle contains zero or more OSC packets and a time tag. The contained packets should be applied at the given time tag.

OscColor

An RGBA color.

OscMessage

An OSC message consists of an address and zero or more arguments. The address should specify an element of your Instrument (or whatever you want to control with OSC) and the arguments are used to set properties of the element to the respective values.

OscMidiMessage

Represents the parts of a Midi message. Mainly used for tunneling midi over a network using the OSC protocol.

OscTime

A time tag in OSC message consists of two 32-bit integers where the first one denotes the number of seconds since 1900-01-01 and the second the fractions of a second. For details on its semantics see http://opensoundcontrol.org/node/3/#timetags

OscTimeError

An error returned by conversions involving OscTime.

Enums

OscError

Represents errors returned by decode or encode.

OscPacket

An osc packet can contain an osc message or a bundle of nested messages which is called osc bundle.

OscType

see OSC Type Tag String: OSC Spec. 1.0 padding: zero bytes (n*4)

Type Definitions

Result